mapper: move common utility functions to mp_utils.cc
