+++ CMakeLists.txt
@@ -626,6 +626,8 @@ if((NOT ANDROID) AND (NOT IOS) AND (NOT CMAKE_SYSTEM_NAME MATCHES "Emscripten"))
         "src/mapper/mp_scrpt.h"
         "src/mapper/mp_targt.cc"
         "src/mapper/mp_targt.h"
+        "src/mapper/mp_utils.cc"
+        "src/mapper/mp_utils.h"
         "src/mapper/mp_text.cc"
         "src/mapper/mp_text.h"
     )
