+++ src/mapper/mp_text.cc
@@ -3,7 +3,7 @@
 #include "color.h"
 #include "input.h"
 #include "kb.h"
-#include "mapper/mapper.h"
+#include "mapper/mp_utils.h"
 #include "mapper/mp_proto.h"
 #include "obj_types.h"
 #include "window_manager_private.h"
