+++ src/mapper/mp_utils.h
@@ -1,0 +1,12 @@
+#ifndef FALLOUT_MAPPER_MP_UTILS_H_
+#define FALLOUT_MAPPER_MP_UTILS_H_
+
+namespace fallout {
+
+void mapperShowTimedMsg(const char* msg);
+bool mapperYesNoDialog(const char* msg);
+void mapperShowMessage(const char* msg);
+
+} // namespace fallout
+
+#endif /* FALLOUT_MAPPER_MP_UTILS_H_ */
